Added search_position, alpha_beta, quiescence, clear_for_search, check_up. Changed types to i32. Added SearchInfo default constructor. Added INFINITE and IS_MATE. Added evaluate. Added search test to main. Updated add_quiet_move functions with killer moves and MVV_LVA. Added generate_all_capture_moves.

+++ src/board.rs
@@ -11,7 +11,7 @@ pub fn check_board(position: &mut Board) {
     let mut temp_big_piece: [u8; 2] = [0; 2];
     let mut temp_major_piece: [u8; 2] = [0; 2];
     let mut temp_minor_piece: [u8; 2] = [0; 2];
-    let mut temp_material: [u16; 2] = [0; 2];
+    let mut temp_material: [i32; 2] = [0; 2];
 
     let mut temp_pawns: [u64; 3] = [0; 3];
     temp_pawns[WHITE] = position.pawns[WHITE];
